requirements.txt
# Список зависимостей проекта
telethon==1.30.3
python-dotenv
Pillow